backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
asyncpg>=0.29.0,<0.30.0
redis>=5.0.0,<6.0.0

langchain>=0.3.31,<0.4.0
langchain-core>=0.3.75,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langgraph>=0.3.31,<0.4.0

httpx>=0.27.0,<0.28.0